Handle rawptr
Not great, but otherwise these guys totally overflow the output;
need to come back to this with java9 where there are better stack
trace APIs.

There is also the option of like.. flagging things? Maybe we can have a
feedback API where, if there are allocations that are unclear, like if
multiple allocations from the allocation tracker matches an elimination,
redo the allocation tracking with detailed regular stack traces on one
of the ambiguous allocation points?

Lots of options.. in any case, the code in this commit is correct
most of the time, but when it's wrong it'll mean users miss allocation
points they perhaps should look into, so this def. needs improvement.