Exclude allocations made by the test harness
Getting the dynamic invocation to not allocate anything turns out to
be tricky - I'd hoped that MethodHandles would be the answer here, but
it turns out to allocate a massive amount the first time it's ran, since
it compiles a little lambda to run the invocation.

We could move to compile something manually, but that all gets very
silly; this solution simply runs an analysis on an empty method, and
then uses that as a reference point for allocations to remove from
reports.