Broken, but a start
I'd very much prefer to avoid separate artifacts for java8/java9;
however, the only way to get stack traces on java8 in a sensible way
is via sun.reflect.Reflect#getCallerClass, and that is not available on
java9. On java9, we want to use the StackWalker API, which isn't
available on java 8.

So; in both cases, we need to use reflection to work with those APIs.
That's all fine and well, except it's in the hottest part of the
allocation tracking code base..

Switching getCallerClass to go via reflection seems to have increased
runtime about 150% (55s -> 85s, give or take), so maybe this is fine?

In any case, there's something wrong with this, because the test
ensuring an empty method yields an empty report is now red.